+++ tdx/converters.py
@@ -8,16 +8,11 @@ from discord.ext import commands
 from dateutil.parser import parse, ParserError
 from redbot.core.i18n import Translator
 import trainerdex
-from tdx.models import Team
+from tdx.models import Faction
 
 _ = Translator("TrainerDex", __file__)
 
-team_search_values = [
-    ["Gray", "Green", "Teamless", "No Team", "Team Harmony"],
-    ["Blue", "Mystic", "Team Mystic"],
-    ["Red", "Valor", "Team Valor"],
-    ["Yellow", "Instinct", "Team Instinct"],
-]
+
 
 
 class TrainerConverter(commands.Converter):
@@ -29,8 +24,6 @@ class TrainerConverter(commands.Converter):
     """
 
     async def convert(self, ctx, argument) -> trainerdex.Trainer:
-        print("TrainerConverter: ", argument, type(argument))
-
         if isinstance(argument, (discord.User, discord.Member)):
             match = None
             mention = argument
@@ -62,18 +55,56 @@ class TrainerConverter(commands.Converter):
             result = trainerdex.Client().get_trainer_from_username(argument)
 
         if result is None:
-            raise commands.BadArgument('Trainer "{}" not found'.format(argument))
+            raise commands.BadArgument(_("Trainer `{}` not found").format(argument))
 
         return result
 
 
 class TeamConverter(commands.Converter):
-    async def convert(self, ctx, argument: str) -> Team:
-        if argument.lower() in [x.lower() for x in team_search_values[0]]:
-            return Team(0, _("Teamless"), "#929292")
-        elif argument.lower() in [x.lower() for x in team_search_values[1]]:
-            return Team(1, _("Mystic"), "#0005ff")
-        elif argument.lower() in [x.lower() for x in team_search_values[2]]:
-            return Team(2, _("Valor"), "#ff0000")
-        elif argument.lower() in [x.lower() for x in team_search_values[3]]:
-            return Team(3, _("Instinct"), "#fff600")
+    def __init__(self):
+        self.teams = [
+            {
+                "id": 0,
+                "verbose_name": _("Teamless"),
+                "colour": 9605778,
+                "lookups": ["Gray", "Green", "Teamless", "No Team", "Team Harmony"],
+            },
+            {
+                "id": 1,
+                "verbose_name": _("Mystic"),
+                "colour": 1535,
+                "lookups": ["Blue", "Mystic", "Team Mystic"],
+            },
+            {
+                "id": 2,
+                "verbose_name": _("Valor"),
+                "colour": 16711680,
+                "lookups": ["Red", "Valor", "Team Valor"],
+            },
+            {
+                "id": 3,
+                "verbose_name": _("Instinct"),
+                "colour": 16774656,
+                "lookups": ["Yellow", "Instinct", "Team Instinct"],
+            },
+        ]
+
+    async def convert(self, ctx, argument: str) -> Faction:
+        if argument.isnumeric():
+            if int(argument) < len(self.teams):
+                result = Faction(**self.teams[int(argument)])
+            else:
+                result = None
+        else:
+            options = [
+                x for x in self.teams if argument.casefold() in map(str.casefold, x.get("lookups"))
+            ]
+            if len(options) == 1:
+                result = Faction(**options[0])
+            else:
+                result = None
+
+        if result is None:
+            raise commands.BadArgument(_("Faction `{}` not found").format(argument))
+
+        return result